requirements.txt
python-dotenv==0.21.0
replicate==0.4.0
discord~=2.1.0
keytotext==2.3.2
PyQRCode~=1.2.1
aiogram~=2.23.1
requests~=2.28.1